requirements.txt
requests
dnspython
colorama